pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mobile-world"
version = "0.1.0"
description = "Mobile GUI automation and testing framework"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "absl-py==2.1.0",
    "android_env==1.2.3",
    "fuzzywuzzy==0.18.0",
    "grpcio-tools",
    "immutabledict==2.0.0",
    "IPython",
    "jsonschema==4.17.3",
    "matplotlib>=3.6.1",
    "numpy>=1.26.3",
    "pandas>=2.2.0",
    "python-Levenshtein",
    "requests",
    "tenacity",
    "termcolor",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "backoff>=2.2.1",
    "jsonlines>=4.0.0",
    "colorama>=0.4.6",
    "xmltodict>=0.15.0",
    "lxml>=6.0.1",
    "loguru>=0.7.3",
    "python-dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "gradio>=5.49.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "opencv-python",
    "python-multipart>=0.0.6",
    "openai>=1.106.1",
    "dashscope>=1.24.6",
    "mcp>=1.9.4",
    "imagehash>=4.3.2",
    "joblib",
    "rich>=13.0.0",
    "qwen-agent>=0.0.31",
    "python-fasthtml>=0.12.33",
    "fastmcp>=2.9.2",
    "openpyxl>=3.0.0",
    "markdownify>=1.2.2",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
agents = ["openai>=1.106.1"]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.6",
    "pylint>=3.0.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "ipdb",
]

all = ["mobile-world[agents,dev]"]

[project.scripts]
mobile-world = "knowu_bench.core.cli:main"
mw = "knowu_bench.core.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/mobile-world"
Repository = "https://github.com/your-org/mobile-world"
Documentation = "https://mobile-world.readthedocs.io"

[tool.hatch.build.targets.wheel]
packages = ["src/knowu_bench"]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade - automatically modernize Python syntax
    "FA", # flake8-future-annotations
]

# Enable specific UP rules for Python 3.12 typing style
extend-select = [
    "UP006", # Use PEP 585 (list[str] instead of List[str])
    "UP007", # Use X | Y for type unions instead of Union[X, Y]
    "UP035", # Import from collections.abc instead of collections
    "UP037", # Remove quotes from type annotations
]

ignore = [
    "E501", # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
# FastHTML uses star imports as its standard pattern
"src/knowu_bench/core/log_viewer/routes.py" = ["F405", "F403"]

[tool.ruff.lint.isort]
known-first-party = ["knowu_bench"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Enable Python 3.12 features
enable_error_code = ["truthy-bool", "redundant-expr"]
